Function definitions in func_declare.h
Significantly reducing number of header files
