+++ wled00/alexa.cpp
@@ -1,9 +1,13 @@
-#include "alexa.h"
 #include "wled.h"
-#include "const.h"
-#include "led.h"
-#include "wled_eeprom.h"
-#include "colors.h"
+
+/*
+ * Alexa Voice On/Off/Brightness/Color Control. Emulates a Philips Hue bridge to Alexa.
+ * 
+ * This was put together from these two excellent projects:
+ * https://github.com/kakopappa/arduino-esp8266-alexa-wemo-switch
+ * https://github.com/probonopd/ESP8266HueEmulator
+ */
+#include "src/dependencies/espalexa/EspalexaDevice.h"
 
 #ifndef WLED_DISABLE_ALEXA
 void onAlexaChange(EspalexaDevice* dev);
