+++ wled00/blynk.cpp
@@ -1,9 +1,9 @@
-#include "blynk.h"
-#include "const.h"
 #include "wled.h"
 #include "src/dependencies/blynk/Blynk/BlynkHandlers.h"
-#include "led.h"
-#include "colors.h"
+
+/*
+ * Remote light control with the free Blynk app
+ */
 
 uint16_t blHue = 0;
 byte blSat = 255;
