+++ wled00/button.cpp
@@ -1,8 +1,4 @@
-#include "button.h"
 #include "wled.h"
-#include "led.h"
-#include "wled_eeprom.h"
-#include "set.h"
 
 /*
  * Physical IO
