--- wled00/button.h
@@ -1,12 +1,0 @@
-#ifndef WLED_BUTTON_H
-#define WLED_BUTTON_H
-#include <Arduino.h>
-/*
- * Physical IO
- */
-
-void shortPressAction();
-void handleButton();
-void handleIO();
-
-#endif // WLED_BUTTON_H
