+++ wled00/file.cpp
@@ -1,7 +1,8 @@
-#include "file.h"
 #include "wled.h"
-#include "led.h"
-#include "notify.h"
+
+/*
+ * Utility for SPIFFS filesystem
+ */
 
 //filesystem
 #ifndef WLED_DISABLE_FILESYSTEM
@@ -11,85 +12,6 @@
 #endif
 #include "SPIFFSEditor.h"
 #endif
-
-enum class AdaState {
-  Header_A,
-  Header_d,
-  Header_a,
-  Header_CountHi,
-  Header_CountLo,
-  Header_CountCheck,
-  Data_Red,
-  Data_Green,
-  Data_Blue
-};
-
-// Maybe Adalight should not be in filehandling? TODO
-void handleSerial()
-{
-  #ifdef WLED_ENABLE_ADALIGHT
-  static auto state = AdaState::Header_A;
-  static uint16_t count = 0;
-  static uint16_t pixel = 0;
-  static byte check = 0x00;
-  static byte red   = 0x00;
-  static byte green = 0x00;
-  
-  while (Serial.available() > 0)
-  {
-    yield();
-    byte next = Serial.read();
-    switch (state) {
-      case AdaState::Header_A:
-        if (next == 'A') state = AdaState::Header_d;
-        break;
-      case AdaState::Header_d:
-        if (next == 'd') state = AdaState::Header_a;
-        else             state = AdaState::Header_A;
-        break;
-      case AdaState::Header_a:
-        if (next == 'a') state = AdaState::Header_CountHi;
-        else             state = AdaState::Header_A;
-        break;
-      case AdaState::Header_CountHi:
-        pixel = 0;
-        count = next * 0x100;
-        check = next;
-        state = AdaState::Header_CountLo;
-        break;
-      case AdaState::Header_CountLo:
-        count += next + 1;
-        check = check ^ next ^ 0x55;
-        state = AdaState::Header_CountCheck;
-        break;
-      case AdaState::Header_CountCheck:
-        if (check == next) state = AdaState::Data_Red;
-        else               state = AdaState::Header_A;
-        break;
-      case AdaState::Data_Red:
-        red   = next;
-        state = AdaState::Data_Green;
-        break;
-      case AdaState::Data_Green:
-        green = next;
-        state = AdaState::Data_Blue;
-        break;
-      case AdaState::Data_Blue:
-        byte blue  = next;
-        setRealtimePixel(pixel++, red, green, blue, 0);
-        if (--count > 0) state = AdaState::Data_Red;
-        else {
-          if (!realtimeMode && bri == 0) strip.setBrightness(briLast);
-          arlsLock(realtimeTimeoutMs, REALTIME_MODE_ADALIGHT);
-
-          strip.show();
-          state = AdaState::Header_A;
-        }
-        break;
-    }
-  }
-  #endif
-}
 
 
 #if !defined WLED_DISABLE_FILESYSTEM && defined WLED_ENABLE_FS_SERVING
