--- wled00/mqtt.h
@@ -1,9 +1,0 @@
-#ifndef WLED_MQTT_H
-#define WLED_MQTT_H
-/*
- * MQTT communication protocol for home automation
- */
-bool initMqtt();
-void publishMqtt();
-
-#endif //WLED_MQTT_H
