+++ wled00/usermod.cpp
@@ -2,7 +2,7 @@
 /*
  * This file allows you to add own functionality to WLED more easily
  * See: https://github.com/Aircoookie/WLED/wiki/Add-own-functionality
- * EEPROM bytes 2750+ are reserved for your custom use case. (if you extend #define EEPSIZE in wled_eeprom.h)
+ * EEPROM bytes 2750+ are reserved for your custom use case. (if you extend #define EEPSIZE in const.h)
  * bytes 2400+ are currently ununsed, but might be used for future wled features
  */
