--- wled00/usermod.h
@@ -1,8 +1,0 @@
-#ifndef WLED_USERMOD_H
-#define WLED_USERMOD_H
-
-void userSetup();
-void userConnected();
-void userLoop();
-
-#endif // WLED_USERMOD_H
